+++ CHANGELOG.md
@@ -21,6 +21,27 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 
+## [0.58.2] - 2025-04-22
+
+### Changes
+
+- increase read timeout limit for BackendApiSink
+
+## [0.58.1] - 2025-04-17
+
+### Changes
+
+- log model info on sink errors
+
+## [0.58.0] - 2025-04-11
+
+### Changes
+
+- move ingest timeout configuration from BackendAPIConnector to BackendApiSink
+- wrap read time out errors of http connector in a custom timed error
+- add proportional backoff to http connector: the longer it took, the longer we chill
+- use watch decorator on sinks to only log once every 1000 write ops
+
 ## [0.57.0] - 2025-04-09
 
 ### Added